+++ Packages/OpenApiCodeGen/Editor/Core/Interfaces/IGenerable.cs
@@ -1,10 +1,11 @@
-using Cysharp.Threading.Tasks;
+using System.Threading.Tasks;
+
 
 namespace ReBeat.OpenApiCodeGen.Core
 {
     interface IGenerable
     {
         ProcessResponse Generate(SettingSchema settingSchema);
-        UniTask<ProcessResponse> GenerateAsync(SettingSchema settingSchema);
+        Task<ProcessResponse> GenerateAsync(SettingSchema settingSchema);
     }
 }