+++ Packages/OpenApiCodeGen/Editor/Presenter/Interfaces/IGenerablePresenter.cs
@@ -8,6 +8,5 @@ namespace ReBeat.OpenApiCodeGen.Presenter
     {
         void Generate();
         void Bind(MenuWindow menuWindow);
-        void OnDestroy();
     }
 }