+++ Packages/OpenApiCodeGen/Editor/Presenter/SetupPresenter.cs
@@ -1,14 +1,5 @@
 #nullable enable
 
-using System.IO;
-
-using Cysharp.Threading.Tasks;
-
-using R3;
-
-using ReBeat.OpenApiCodeGen.Core;
-using ReBeat.OpenApiCodeGen.Dto;
-using ReBeat.OpenApiCodeGen.Lib;
 using ReBeat.OpenApiCodeGen.Model;
 
 namespace ReBeat.OpenApiCodeGen.Presenter
@@ -25,18 +16,13 @@ namespace ReBeat.OpenApiCodeGen.Presenter
         public void Bind(SetupMenu setupMenu)
         {
             _setupMenu = setupMenu;
-            _setupMenu.SetOnChangeHandler((dto) => _setupModel.SetupMenuDto.Value = dto);
-            _setupModel.Status.Subscribe(s => _setupMenu.SetProgressStatus(s));
+            _setupMenu.SetOnChangeHandler((dto) => _setupModel.SetupMenuDto = dto);
+            _setupModel.OnChangeStatus += (s) => _setupMenu.SetProgressStatus(s);
         }
 
         public bool CheckRunnableDockerPath()
         {
             return _setupModel.CheckRunnableDockerPath();
-        }
-
-        public void OnDestroy()
-        {
-            _setupModel.Dispose();
         }
 
         public void Setup()