+++ Packages/OpenApiCodeGen/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.2]
+
+### Fixed
+
+- Change to EnumField at OpenAPI Library Setting
+- Change to independent R3, UniTask, .NET.Json
+
 ## [0.2.1]
 
 ### Fixed